+++ todomanager/src/main/kotlin/com/github/kokorin/todo/dao/TodoListDao.kt
@@ -6,11 +6,11 @@ import com.github.kokorin.todo.model.TodoList
 interface TodoListDao {
     fun getAllTodos(): List<Pair<TodoList, List<Todo>>>
 
-    fun addTodo(name: String, description: String, listId: Long)
+    fun addTodo(name: String, description: String, listId: Int)
 
     fun addTodoList(todoListName: String, todoListDescription: String)
 
-    fun removeTodoList(todoListId: Long)
+    fun removeTodoList(todoListId: Int)
 
-    fun markTodoAsDone(todoId: Long)
+    fun markTodoAsDone(todoId: Int)
 }
